+++ lib/decorators/src/public_api.ts
@@ -3,3 +3,4 @@ export { Persistence } from './persistence/persistence';
 export { StateRepository } from './state-repository/state-repository';
 export { debounce } from './debounce-time/debounce';
 export { payload } from './payload/payload';
+export { named } from './named/named';